# cmd/gateway/openapi.yaml
openapi: 3.0.3
info:
  title: AI Gateway
  version: 1.7.0
  description: |
    Unified AI Gateway — OpenAI / Anthropic / Responses under one domain.
    Gateway auth: Bearer sk-gw-* or admin session JWT on /v1/*.
    Admin: Bearer JWT on /api/*. Errors: {error:{message,type}}.
servers:
  - url: http://localhost:8080
    description: local
  - url: https://ai.example.com
    description: production
paths:
...
  /api/models/catalog:
    get:
      summary: List models catalog
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: provider
          in: query
          schema: { type: string }
        - name: reasoning
          in: query
          schema: { type: boolean }
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        '200': { description: catalog }
...
  /api/profile:
    get:
      summary: Current dashboard user profile
      tags: [admin]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: profile }
    put:
      summary: Update current user display name
      tags: [admin]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                display_name: { type: string }
      responses:
        '200': { description: updated profile }
  /api/profile/password:
    post:
      summary: Change own password (revokes other sessions via token_version bump)
      tags: [admin]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                old_password: { type: string }
                new_password: { type: string }
                password: { type: string, description: legacy alias for old_password }
      responses:
        '200': { description: password changed }
        '400': { description: old/new password required }
        '401': { description: old password incorrect }
  /api/profile/activity:
    get:
      summary: Recent request activity for the caller
      tags: [admin]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: activity rows }
  /api/profile/logins:
    get:
      summary: Recent login history for the caller
      tags: [admin]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: login rows }
...
  /api/providers/{id}/discover:
    post:
      summary: Run model discovery against one provider's /models endpoint (admin only, audited)
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: discovered model ids
          content:
            application/json:
              schema:
                type: object
                properties:
                  discovered: { type: array, items: { type: string } }
        '404': { description: provider not found }
  /api/discover-all:
    post:
      summary: Run discovery across all providers (admin only, audited)
      tags: [admin]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: discovery report }
...
  /api/logs:
    get:
      summary: Recent request logs (latest 100)
      tags: [admin]
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: logs }
  /api/logs/{id}:
    get:
      summary: Single request log incl. stored bodies (org-scoped)
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: log detail }
        '404': { description: not found }
...